Refactoring: s/queries/query/g
